# internal/controller/docs/swagger.yaml
definitions:
...
  models.LoginResponse:
    properties:
      authorize:
        type: string
      user_id:
        type: string
    type: object
...
  models.RegisterRequest:
    properties:
      password:
        type: string
      phone_number:
        type: string
    type: object
  models.StandardErrorModel:
    properties:
      error:
        type: string
    type: object
...
info:
  contact:
    name: Soburjon
  description: This is an auto-generated API Docs.
  termsOfService: http://swagger.io/terms/
  title: API
  version: "1"
paths:
...
  /register/login/:
    post:
      consumes:
      - application/json
      description: user id va authorize qaytaradi
      parameters:
      - description: login
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/models.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      summary: user id va authorize qaytaradi
      tags:
      - register
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"